[fix] Fixed String conversion on error retrieval
The error response from the server can be retrieved either in Map or in
String. Previously, in ´getErrorFromResponse` method it was not checked
whether a response is retrieved in String or in Map data type.
Therefore, when the error is retrieved in String format, it would try to
convert String into Map, which is not possible since String does not
have key value and therefore a `MismatchedInputException` from
`com.fasterxml.jackson` library would be thrown.

This change includes the guard for checking in which data type (format)
the error was retrieved in and appropriately handles handles the error
message based on the data type.

Closes #69