[fix] Fix implementation of ´getResponses` method
Previous implementation of `getResponses` method ignored returned Map
from `hostControl_getActiveSessions` RPC method. This lead to, for data
retrieved from `hostControl_getActiveSessions` RPC method to not be
displayed when session information is requested.

This commit fixes this issue, by not ignoring the fact that the data
returned from `hostControl_getActiveSessions` RPC method is in Map
format and it reads the returned data correctly.

Closes #70