[fix] Fix MFA new password dialog not displaying
It is possible, that ´displayMFADialog´ is called from EventDispatchThread
or any other thread. Previously, when method ´displayMFADialog´ was
called from EventDispatchThread, the error would occur due to calling
`SwingUtilities.invokeAndWait` method in EventDisptachThread.

This commit adds a guard, that checks if the current thread is
EventDisptachThread, if so it does not call
`SwingUtilities.invokeAndWait` method, it simply calls `setVisible`
method for the dialog and vice versa.