[fix] With SHIFT current selection underlining
When user navigates client tree using SHIFT + UP/DOWN keys, the
currently selected client leading up from the anchor selection (first
selection) would be underlined. However the currently selected client
leading down from the anchor selection would not be underlined.

This issues occurs due to the way JTree handles selection. When
selecting nodes in a JTree with SHIFT + UP/DOWN keys, the selected
node's path is not immediately added to the selection paths. It is only
added after user selects another node. Therefore, when calling
tree.getSelectionPath(), one would retrieve the previously selected
node's path and not the current.

However, calling tree.getLeadSelectionPath() returns a path, that is
considered a lead path. In this case the lead path is currently selected
node's path. Though, it is not always the case, that the lead path is
currently selected node's path.