[rem] Remove deleting client specifc values option
The option for deleting client specifc values and setting client values
to server defaults for product properties, returned the same results.
Even though there were some small differences as to how they operated.

To remove the option for deleting client specific values instead of
setting client values to server defaults was made because the former was
somewhat buggy in how it worked and the latter made more sense based on
the description.