[fix] Select currently selected value
In previous fix (73814db3) there was an unforeseen error. The fix worked
as it should, besides of still having the original issue (only with
extra step). The error occured, whenever a user would set products'
properties to a server defaults.

This commit fixes both the original issue and the issue, that previous
fix (73814db3) did not fix. It fixes by selecting in possible values for
property dialog, the currently selected properties' value for a
property in properties table. This fix also works, if a user sets
products' properties to a server defaults.

Closes #77