[chg] Retrieve all products; Separate later
When data is loaded for the dashboard, the product data (idents) will be
retrieved by executing `productOnDepot_getIdents` without applying
filter for the product type. This way all products (localboot and netboot)
will be retrieved. The retrieved products will then be processed as
required. This reduces the calling of `productOnDepot_getIdents` RPC
method and improves the data load performance by 2 seconds.