[chg] Update client products' states when viewed
This makes sure that updates, that are triggered by messagebus, will
only happen, when the products view is currently opened for the specific
client. It also reduces the `productPropertyState_getObjects` RPC method
calling, since the updates are applied for currently selected client
only.

If more than one client is selected and product view is opened, the
updates will be ignored for all clients by messagebus.