[chg] Do nothing if selection hasn't changed
The `valueChanged` method is triggered even though the original
selection in client table hasn't changed. This leads, to multiple RPC
method calls for the following methods:

- productOnClient_getHashes (4 times)
- productPropertyState_getObjects (4 times)

whenever the global reload function is used and when a user views the
localboot or netboot product page during the reload.

To prevent this we ignore the trigger of `valueChanged` method if the
original selection in client table hasn't changed and also if the
current selection is empty (this prevents calling multiple RPC
methods when selection is temporarily cleared).

With current changes the RPC method calling is reduced to the following:

- productOnClient_getHashes (1 time)
- productPropertyState_getObjects (2 times)

`productPropertyState_getObjects` RPC method is called two times with
different parameters. It retrieves product property states for all
depots and currently selected clients.