[fix] Provide workaround for seven zip lib init
- Issue

The Seven Zip library creates a 7-Zip-JBinding.dll file inside the
SevenZipJBinding-<HashValue> directory, where <HashValue> is generated
automatically by the library itself. The directory itself is located in
platform's temporary directory.

Theoretically, the generated <HashValue>, that is appended to the
SevenZipJBinding directory name, should be different each time, but
it is not. The library contains an incorrect conversion of a hash value
Which leads to a problem of java process accessing the library file in
the same location. When one java process tries to access the library
file that is used by another java process it is unable to access it and
the exception is thrown.

The issue in library of incorrect hash conversion lead to, not being
able to open multiple ZIP files in opsi-logviwer through Windows
Explorer context menu. Because the second opsi-logviewer process, that
opens a second ZIP file, tries to access the library file, that is
already used by the first opsi-logviewer process.

- Solution (Workaround)

The error in the libraries code already contains a fix, though that fix
is not yet published (the library has not been updated since 19. February
2020; Fix was implemented in 5. July 2022).

Pull request that fixes the issue: https://github.com/borisbrodski/sevenzipjbinding/pull/34.

Because there is no official release of the library with current fix,
a simple workaround was implemented. The workaround implementation
creates an extra directory inside the platform's temporary directory and
specifies the new directory for Seven Zip library, where the library
file should be saved. This makes sure that all java process access
library file from different locations.