[chg] Suppress dead code warning
Dead code warning occurs due to usage of global variable THEMES. By
default THEMES is set to false, indicating that themes feature is
disabled. The code inside the if statement is reliant on this feature to
be enabled, therefore when the THEMES is false the code doesn't get
executed and since THEMES is not dynamic - doesn't change in runtime -
we get the dead code warning.