[fix] Fix client tree selection
Issue

When the same client in client tree was selected multiple times in a
row and afterwards the new client was selected, the data was not loaded
for the newly selected client

This happened due to incorrect provided information by the
`getSelectedClients()` method, since it has not been updated with the
new selection. The `getSelectedClients()` method returned previously
selected clients and not currently selected clients. This lead to
`previousSelectedClients` be the same as a result of
`getSelectedClients()` method, which meant that `valueChanged()` method
triggered by the new selection was ignored.

Solution

We now use `getSelectedValues()` method available in
`JTableSelectionPanel` to retrieve currently selected clients.