[fix] Do not override getColumnClass method
- Issue

The implementation of `getColumnClass` method returned null for column
class, when the value in the first row of the column was null. This lead
to, being `NullPointerException` be thrown in Localboot/Netboot product
tabs, when first row of the column had null value (empty value).

- Solution

We simply remove `getColumnClass` implementation in
`InstallationStateTableModel` class, since there is no need to provide
specific logic, when default does the job.

The comment (above `getColumnClass` method implementation) is no longer
valid.