[ref] Use retrieve methods to retrieve PD
- Create retrieve methods for get methods, that are associated with
  persistent data.
- Use retrieve instead of get methods in `ReloadHandler` implementations
  to retrieve new data from the server.

The idea behind this change is `Separation of Concerns`. The retrieve
methods are responsible for retrieving new data from the server and
saving it internally. The get methods are responsible for retrieving
internally saved data.

The get methods, that do not deal with persistent data, do not have
(require to have) retrieve methods. However, if they do, the retrieve
methods should be declared private.