[fix] Display version for once/custom products
This commit fixes the issue of not displaying product version, for
products, that have been installed as once or custom installation, even
if the product has version.

The issue happend due to `installationStatus` not changing, which can be
seen when calling `productOnClient_getObjects` on client that has
products installed with once or custom installation. The
`installationStatus` does not change even though product is installed
with once or custom installation is because once installation sets
`installationStatus` of a product to `not_installed` and custom
installation has no affect on `installationStatus` of a product.