[fix] Double values added when editing properties
This commit fixes issue of double values added when editing a property,
that can have multiple values. This issue arises due to incorrect
implementation of a for loop, that handles addition of multiple values
(this bug was introduced in commit 6136e1f7c37d9c9761aac756036fc2460732b414).

In the commit 6136e1f7c37d9c9761aac756036fc2460732b414 it was done so
that the last element would be added twice.

To fix this issue we revert to previous implementation prior to commit
6136e1f7c37d9c9761aac756036fc2460732b414.