[fix] Client deletion on opsi 4.2 servers
the RPC method `host_delete` on opsi 4.2 servers does not support a list
as a parameter, which results in an error:

Opsi service errror: [DataError] Invalid input of type: 'list'. Convert
to a bytes, string, int or float first.

However, in opsi 4.3 servers the RPC method `host_delete` does support a
list as a parameter.