[rem] Temporarily remove modules file upload
Currently unable to implement due to the absence of a suitable method
for uploading module file. Both existing file upload methods, WebDAV and
Messagebus, are not viable options:

- WebDAV: Doesn't have the required directory enabled/activated for
  module file upload to work. Also, there is no mechanism available to
  enable/activate the required directory.
- Messagebus: `file_upload` events require opening a terminal on the
  server, to upload files.

The module file upload functionality will be reimplemented once a
suitable method for uploading module file is available, which
necessitates backend work.