[fix] Check permissions in UserRolesConfigDataService
If no module is present for user-roles, we ignore user-role
configuration.

Previously some feature activations where managed in `UserConfig`, which
lead to those features being deactivated/activated without the presence
of a module for user-roles. We now check permissions of all features in
`UserRolesConfigDataServer`, which makes sure that user-roles only work
then, when module is present.